refactor: co-locate OG service layers
