+++ apps/web/src/pages/og/[...slug].png.ts
@@ -2,8 +2,10 @@ import type { APIRoute } from "astro"
 import * as Layer from "effect/Layer"
 import * as FetchHttpClient from "effect/unstable/http/FetchHttpClient"
 import { HttpRouter, HttpServer } from "effect/unstable/http"
-import { OpenGraphRoute } from "@/features/open-graph/Http"
-import { OpenGraphMiddlewareLive } from "@/features/open-graph/Live"
+import {
+  layer as openGraphLayer,
+  route as openGraphRoute,
+} from "@/features/open-graph/Http"
 
 // On-demand server endpoint: slugs derive from arbitrary page pathnames
 // (see BaseLayout.getOgImagePath), so the route cannot be enumerated at build
@@ -11,9 +13,7 @@ import { OpenGraphMiddlewareLive } from "@/features/open-graph/Live"
 export const prerender = false
 
 const AppLayer = HttpRouter.layer.pipe(
-  Layer.provideMerge(
-    OpenGraphRoute.pipe(Layer.provide(OpenGraphMiddlewareLive)),
-  ),
+  Layer.provideMerge(openGraphRoute.pipe(Layer.provide(openGraphLayer))),
   Layer.provide(FetchHttpClient.layer),
   Layer.provide(HttpServer.layerServices),
 )
