+++ apps/web/test/stubs/astro-assets.ts
@@ -1,0 +1,3 @@
+export const fontData = {}
+
+export const experimental_getFontFileURL = (url: string): string => url
