+++ apps/web/test/stubs/astro-content.ts
@@ -1,0 +1,2 @@
+export const getCollection = (): Promise<never> =>
+  Promise.reject(new Error("astro:content is unavailable in unit tests"))
